// packages/cmk-frontend-vue/src/dashboard/components/DashboardMenuHeader/types.ts
/**
 * Copyright (C) 2025 Checkmk GmbH - License: GNU General Public License v2
 * This file is part of Checkmk (https://checkmk.com). It is subject to the terms and
 * conditions defined in the file COPYING, which is part of this source code package.
 */
import type { DashboardOwnerType } from '@/dashboard/types/dashboard'

export interface SelectedDashboard {
  name: string
  owner: string
  title: string
  type: DashboardOwnerType
}